requirements.txt
Flask==1.1.2
pandas==1.2.3
gunicorn==19.6.0
requests==2.22.0
pyarrow==3.0.0
google-cloud-bigquery
google-cloud-storage